chat-service/requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.3
pydantic
python-dotenv~=1.0.1
sqlalchemy[asyncio]~=2.0.31
aiomysql
redis~=5.0.7
httpx
Faker~=26.0.0
mysql-connector-python~=9.0.0
starlette~=0.26.1
PyJWT~=2.8.0
cryptography
